use numeric field for cart form and fix recursion issue

+++ src/admin/PriceField.php
@@ -12,15 +12,16 @@ use SilverStripe\Forms\CurrencyField;
  * @package swipestripe
  * @subpackage admin
  */
-class PriceField extends CurrencyField {
+class PriceField extends CurrencyField
+{
 
 	/**
 	 * Render field with custom template
 	 * 
 	 * @param array $properties
 	 */
-	public function FieldHolder($properties = array()) {
-
+	public function FieldHolder($properties = array())
+	{
 		$shopConfig = ShopConfig::current_shop_config();
 		$properties = array_merge($properties, array(
 			'BaseCurrency' => $shopConfig->BaseCurrency
@@ -35,25 +36,32 @@ class PriceField extends CurrencyField {
 	 * 
 	 * @param mixed $val
 	 */
-	public function setValue($val, $data = null) {
-		if(!$val) $val = 0.00;
+	public function setValue($val, $data = null)
+	{
+		if (!$val) $val = 0.00;
 		$shopConfig = ShopConfig::current_shop_config();
 		$precision = $shopConfig->BaseCurrencyPrecision;
 
-		$this->value = number_format((double)preg_replace('/[^0-9.\-]/', '', $val), $precision);
+		$this->value = number_format((float)preg_replace('/[^0-9.\-]/', '', $val), $precision);
 		return $this;
 	}
 
-	public function validate($validator) {
-		if(!empty ($this->value)
-				//validate against any number of digits after the decimal place
-				&& !preg_match('/^\s*(\-?\$?|\$\-?)?(\d{1,3}(\,\d{3})*|(\d+))(\.\d+)?\s*$/', $this->value)) {
+	public function validate($validator)
+	{
+		if (
+			!empty($this->value)
+			//validate against any number of digits after the decimal place
+			&& !preg_match('/^\s*(\-?\$?|\$\-?)?(\d{1,3}(\,\d{3})*|(\d+))(\.\d+)?\s*$/', $this->value)
+		) {
 
-			$validator->validationError($this->name, _t('Form.VALIDCURRENCY', "Please enter a valid currency"),
-				"validation", false);
+			$validator->validationError(
+				$this->name,
+				_t('Form.VALIDCURRENCY', "Please enter a valid currency"),
+				"validation",
+				false
+			);
 			return false;
 		}
 		return true;
 	}
-
 }
